+++ packages/custom-function-parser/test/get-argument-type/0001.mjs
@@ -17,6 +17,20 @@ import { parse } from '@csstools/custom-function-parser';
 }
 
 {
+	const fn = parse('--foo(--bar, --baz type(<string>): "foo")');
+
+	assert.strictEqual(
+		fn.parameters[0].getArgumentType(),
+		'',
+	);
+
+	assert.strictEqual(
+		fn.parameters[1].getArgumentType(),
+		'<string>',
+	);
+}
+
+{
 	const fn = parse('--foo(--bar /* a comment */ <string> /* a comment */ : /* a comment */ "foo")');
 
 	assert.strictEqual(
@@ -26,7 +40,25 @@ import { parse } from '@csstools/custom-function-parser';
 }
 
 {
+	const fn = parse('--foo(--bar type(/* a comment */ <string> /* a comment */) : /* a comment */ "foo")');
+
+	assert.strictEqual(
+		fn.parameters[0].getArgumentType(),
+		'<string>',
+	);
+}
+
+{
 	const fn = parse('--foo(--bar [ /* a comment */ <string> /* a comment */ | /* a comment */ <color> /* a comment */ ] : /* a comment */ "foo")');
+
+	assert.strictEqual(
+		fn.parameters[0].getArgumentType(),
+		'[  <string>  |  <color>  ]',
+	);
+}
+
+{
+	const fn = parse('--foo(--bar type([ /* a comment */ <string> /* a comment */ | /* a comment */ <color> /* a comment */ ]) : /* a comment */ "foo")');
 
 	assert.strictEqual(
 		fn.parameters[0].getArgumentType(),